Set up Cypress for tests
Install Cypress as a dependency and add some npm scripts for running it
more conveniently. I added some simple tests for the placeholder page
that Express is serving right now just so we have something to run. I
have a feeling I'm going to need to tweak this set up, because the way
we run the integration tests and start up both cypress and the dev
server feels a little kludgy to me.