Replace Nunjucks with Svelte
Since the routing for this app should be relatively simple, I think I
can set up server-side rendering on my own using Svelte for an
isomorphic app. I'm not sure if this is a great idea or if it would be
more performant to just stick with vanilla HTML/CSS/JS and send HTML
fragments over the wire, but the developer experience with Svelte is
pretty great.