Render a map of the states and counties
This is currently an SVG, which has the advantage of being something I
can render server-side and ship along with the rest of the HTML. The
downside is that the path is a bit large, so it's slow. It might be
better to do this as a web component that draws to a canvas. That way I
know how big the image needs to be and can calculate simplified paths
that will look good at that resolution.