Draw the borders on the basemap
I've got a visible US map on the page now that is scaled to fit the
window. Resizing is really slow, though. So I need to look into some
strategies to speed things up. I think I'll want to throttle the redraw,
and I may even want to consider doing more of the projecting and such on
the server, since all of the geo stuff is really increasing the size of
my bundles.