Prevent bounding box from clipping the souther USA
I added midpoints to the polygon defining the bounding box because the
Albers USA projection I'm using curves downwards. Midpoints help D3 keep
the southern parts of the USA (such as Texas and Florida) in the
viewport.

I also removed the boundning box drawing now that this is sorted enough
to move on.