Add build steps to package.json
The Dockerfile needs an npm build script, so here one is. I had to move
a number of dependencies out of devDependencies for this so that we
don't waste time installing cypress and such. I should look into other
ways of defining dependencies so that we don't have to have something
like Rollup installed in the production container.