Move static files to a public/ directory
Move the location of all the static files that we host out of
src/static/ and into public/. public/ is more conventional for Express.

I've also added a task to copy the font files out of USWDS instead of
keeping them in Git. I think this makes more sense as it should be
easier to update. If USWDS changes their fonts, we probably will, too.