Remove all of the files in src/static/
This directory was a problem because it mixed built files (which we
ignored) like CSS and bundled JS with files we committed like the font
files. Now that we're managing all that with npm scripts, it's much
simpler to just remove this and have a plain .gitignore in the project
root.