Implement Show Counties toggle
Add another component containing the controls for the map. I've made
this part of the map so that it's easy to bind the showCounties property
from the HrrrMap to the HrrrControls without needing to involve a store.
The map checks the search parameters in the URL for the showCounties
parameter to determine the initial state of the checkbox.