Add plots for each of the smoke bands
Just as a test to make sure I can get to the data, I plotted each of the
bands in the GRIB file containing smoke density values. This at least
shows that there's something in there and seems to confirm that each
band is a different slice of the atmosphere.