Write out a JSON file for use with the frontend
Now that I've got a workable cross-section chart made with matplotlib
using Jebb's code, I can take the data that bult that chart and write it
out as a JSON file that I can use in the frontend to start working on
the UI. The JSON file is 2MB for 1,200 steps, which we can compress down
to 812KB with gzip. It's big, but it doesn't seem unreasonable.

Worst case scenario: we may consider using d3-contour in a lambda
function to generate the contours and pass those to the client for
rendering. But I think a 2MB data file is ok for this.