Generate a preliminary cross-section
It's upside down, and either the scale or the thresholds aren't quite
right (possibly both), but this does look like it could actually be the
same visualization that we've generated in Python.

I removed the HRRRMap component from the application, since I'm not
planning on using that for now, I just want to focus on getting the
cross-section drawn from this sample JSON file so I can work on getting
a version of the app that just takes the start- and end-coordinates and
gives you a cross-section. We can worry about overviews and insets
later.