Create a rudimentary resizable chart
I bound the viewBox of the SVG in the XSection component to the width
and height of the containing div so that it will resize when the window
resizes. It doesn't currently shrink vertically, so I need to come up
with a better overall layout for the application on larger windows
without causing loss of functionality on smaller viewports.

Now that I'm getting the width and height for the viewBox, though, I
should be able to set up some transformations for the paths of the
contours so that they render right-side up.