Add a left axis to the chart
The AxisLeft component is modeled after the axes generated by d3-axis.

There's a name conflict in our CSS classes between the axis labels
(which are handled in HTML, rather than SVG) and the axis itself. This
isn't causing a problem because of Svelte's scoped styles, but it might
be worth changing. I think it makes sense to leave these as scoped
styles since they aren't generic, but rather apply only to axis labels
in this particular component with its particular markup.