Switch to an official Python base image
The Amazon Lambda containers are too stripped down to easily install
Proj 8.1, which we require for Cartopy 0.20. For now, we're having to
build Proj from source because neither the latest Debian nor the latest
Alpine have Proj 8.1 in their package managers (they're both still on
7.2). Hopefully soon Proj 8.1 will end up in a stable branch of Alpine
(it's currently in edge) and then we can switch to python:3.9.7-alpine
and just `apk add proj` instead of buildng it ourselves, which is quite
slow.