Remove the old frontend service
The development environment I'd set up to accommodate this homegrown
pre-rendered HTML was a little too flaky. It would sometimes need to be
restarted to keep it working. Since we're not really gaining anything
with the pre-rendered HTML anyway (this application can't function
without JavaScript) it's probably not much of a loss.

This kind of application seems like a pretty good candidate for a
single-page app anyway, since it is more like an application with a
single view that updates different parts of the page.