Initialize the Svelte template
Adding all of the files from the default Svelte template created with
`npm init svelte`. I'm committing everything from the template first
just in case we decide we want to revert back to some of the default
configuration, but I'm going to start clearing out a bunch of stuff
next.