Remove Cartopy
I successfully eliminated Cartopy from the dependencies. I was only
using it to get a pyproj.Geod object anyway, so I'm just using pyproj
directly. The odd thing is that the distance that the pyproj Geod
returns is about 15m shorter than the distance reported by the pyproj
Geod that Cartopy gave me. I'm not sure why. I wonder if Cartopy had
some additional CRS information that it passed on to pyproj that I
wasn't seeing.

Having gone to all of that trouble, I shaved about 90MB off of the image
size and eliminated none of the vulnerabilities found by the ECR scan.
Maybe a stock image is not the way to go.