Revert to installing all MetPy dependencies
I must've messed something up with Docker yesterday because I thought
the installation without matplotlib was working, but today I'm getting
import errors because matplotlib is missing. I'm not sure it's saving me
all that much anyway, because a lot of the security errors are related
to the base image. I'll likely leave this for now because it's a
prototype, and in the future I'll look for a more secure image to reduce
some of these errors.

I have removed Cartopy, though, and replaced it with direct use of
pyproj.Geod instead.