Ignore the local cache created by the Docker container
To avoid permissions errors caused by setting the user to the host's
UID, the Docker container running the frontend moves the npm cache to
the WORKDIR, which has the local directory bound over it, so the cache
ends up in our working tree. Since we don't want an npm cache committed
to Git, we ignore it.