Bind controls to Svelte stores to fetch xsections
I set up several writable stores for the cross-section settings and
created a derived store that should fetch the cross-section data from
the Lambda function. Considering, however, the amount of effort that
goes into computing these cross-sections, it might be wise to have a
button for users to explicitly request new data to avoid needing to
debounce the xsection store and cancel running requests.

The fetch is failing at the moment because it can't resolve the hostname
of the backend, so I need to review how name resolution and networking
work with Docker Compose.