Bump xarray version
This happened when I was debugging some errors with the Lambda function
that were caused by the fact that the dependencies are installed into
/var/task/, which I was overwriting with my bind mount from the dev
environment. I ended up removing the bind mount so this wouldn't be a
problem, even though that makes development less easy. But since
everything is working with the new xarray, it's worth keeping.