Trigger updates manually
Because of the amount of work involved in these computations, I think
it's better for the user to signal when they're done entering the
configuration instead of reactively updating as they enter their path
variables. So I've changed the path definition to a single object that
can be set all at once by the controls in response to the user clicking
an update button.