Refactor the USWDS form controls
Since I can't dynamically bind the type when I create a text input, I'll
need different Svelte components for different input types. So to save
on typing, I created a FormGroup component with a slot for the input.
For now it just handles labelling the input, but I imagine it will take
on some additional display responsibilities for errors and help text.