Style the sidebar
Create the appearance of a single sidebar in the application by coloring
the HRRR controls dark. Given the layout of the application, I don't
think having a gray background is necessary, so I switched it to white.
We're not really going to have distinct sections of content surrounded
by gray, it's just the dark side bar and the charts and maps.