Add vertical spacing to the coordinate pairs
Add some margin to the fields in the coordinate pair components so that
the constituent fields are separated a little from the fieldset legend.
For some reason, row-gap isn't working here and I can't figure out why,
so I used the lobotomized owl selector to apply margin-block-start to
the children of the component.