Convert project structure to a Flask app
I'm going to see about running everything off of a GSL VM and providing
the data from a Flask app. I've added the basic structure for a Flask
app and moved the Lambda function handler into a module called routes
in the hrrr_smoke package. It's still a Lambda function for the moment,
going to set up packaging for the dependencies first, then update it for
Flask.