Convert the Lambda function to a Flask route
Converted the handler for the Lambda function to a regular Flask route
in a blueprint. I had to add s3fs as a dependency for now just to avoid
having to deal with accessing GRIB files locally. I think that will be
in a subsequent branch. For now the goal is to just get what was a
Lambda function converted to a Flask app.