Update the frontend to fetch from the new API
I am allowing CORS from all origins for now, though I'll want to switch
this to be only on in the dev environment. I may look into the
flask-cors plugin for this, or I may just add my own decorator. The
frontend now fetches the data from the Flask app, though. So progress!