Remove /xsection/ from HRRR API env variable
Now that we're working with multiple endpoints for fetching things like
the available forecasts, the URL for the API needs to be more generic,
so I've removed the /xsection/ part of the environment variable and
added that to the fetch call in our store. I exported the constant from
the store module as well, so that I can use it in components to fetch
data that doesn't need to be globally available, such as the list of
forecasts.