Add a dropdown component
Forecast selectoin will require two selects: one to choose the forecast
run time, and one to choose the forecast hour. So I've encapsulated
select controls in a dropdown. They're still a little rough around the
edges. I'm unsure if I should provide a placeholder the way the USWDS
components do, or not. It'll take some work to get the reset behavior
working in Svelte with a placeholder.