Implement forecast selection for HRRR controls
Added two dropdowns to the HRRR controls for forecast selection. I'm
thinking it might be best to move these off to a separate component to
keep the fetching of the forecasts isolated from the rest of the
controls and only expose the selected forecast so that the controls can
set the appropriate variable in the store when "Update" is clicked.