Add forecast parameter to fetch requests
I created a writable store for the forecast and bound it to the output
of the forecast hour select. The xsection store now derives from both
the forecast and the path stores, so it will update once everything's in
place, although I think using stores this way is a little wonky. Maybe
this store should refresh on an event, rather than having to have guards
to stop the reactivity. It's not really reactive, since I only want it
to fetch when the user clicks the update button.