Generate cross-section from GRIB files
Instead of relying on existing Zarr files, we'll read directly from GRIB
files, which can be mounted on the VM inside of NOAA's network. Although
this is pretty slow right now, we can probably implement some kind of
caching of the datasets to speed things up. But for now, at least, we
should be able to deploy this to a VM and have it load live data.