Use app config instead of environment variables
I forgot that I was loading the environment variables into the Flask app
config when I updated the xsection route to read from the GRIB files.
Outside of the app configuration, nothing should be reading from the
environment, this way we only have to update config changes in one place
(for example, if we rename an environment variable).