Create a CLI to convert GRIB to Zarr
Since I've had trouble getting responses to return on the server when
reading directly from the GRIB2 files, I've created a new command-line
interface for converting the GRIB2 files to Zarr arrays. Hopefully the
backend can read the Zarr arrays and perform the cross-section without
running out of memory and getting killed. The CLI is something we can
set up with either a watch command using inotify, or using cron to
update the Zarr arrays.