Store each forecast in a separate group
For now, to keep the amount of storage we use lower, we'll only store
the most recent forecast. So instead of choosing an analysis date and a
valid time, users will pick a valid time and they'll get whatever the
most recent analysis produced a forecast for that time. Each forecast is
stored in a separate group to avoid having to determine whether or not
we need to append a new dimension or overwrite a previous forecast.
Since we are, for the moment, only displaying a single forecast hour at
a time, this shouldn't pose a problem.