Read data from the Zarr array instead of the GRIB files
Now that I've got a CLI for pre-processing the GRIB files into Zarr
arrays, I can read from that instead of always having to read the GRIB
files. This seems to be much faster, so clearly a lot of the processing
time on the previous iteration was in reading GRIB files.