Refactor the API fetch calls into a module
This centralizes knowledge about our URL paths and hides that knowledge
behind some function calls, which is nicer for the code. I was also able
to eliminate the checks for the derived cross-section data. Instead,
HrrrControls executes the fetch when "Update" is clicked and sets it to
the xsection store, which is now just a writable store. It simplifies
a lot of code flows.