Start a shell script for generating the map borders
I deleted the Makefile that produced the borders needed for the map, but
we still need a way to reproduce the TopoJSON used for the map in the
frontend. This script will process a shapefile into TopoJSOn and
simplify it to keep the file sizes down. Finding and downloading the
shapefiles will remain an exercise for the developer.