Draw the inset map centered on the path
I've updated the HrrrMap component to react to the path store and draw
the county boundaries that are near the selected cross-section path.
There's a lot of styling and adjustments I'll need to make to this in
order to make it look right (right now it's so scaled down as to be
barely visible), but the fact that it responds to the path and zooms in
correctly is big.