Add the html, body height reset
Learned this recently from Josh W. Comeau. Set the height of the html
and body elements to 100% so that heights of 100% will work more as
expected. This may be part of the solution we need to get things to
size like this is an application window instead of a document.

https://www.joshwcomeau.com/css/custom-css-reset/