Offset the path line from the geo borders
Draw a thicker line in the background color first, then draw the path
over it to create a little space around the line and increase its
contrast against the geographical boundaries. Preventing the borders
from intersecting with the path at all seems to make it a little easier
to read.