Redraw the inset map on resize
I think there was a race condition or something that occurred when the
map was resized. If the drawing occurred before the HTML element was
resized, the drawing was getting erased (which seemed to always happen).
By moving the drawing code into afterUpdate and adding a guard to ensure
we don't draw anything if we're not ready, the problem is fixed. Now the
map will redraw whenever the window is resized.