Update browser history when settings change
Whenever the settings for the cross-section change, update the browser
history with pushState. This will enable the use of forward/back buttons
in the browser, and facilitate sharing of charts by allowing users to
copy the current location from the location bar.

I had to move the initialization code for the stores into main.js
instead of when the App component is mounted because the binding in
App.svelte gets triggered before it is mounted, and was causing the URL
parameters to be cleared out.