Add a clean command to remove old forecasts
In an effort to manage our storage use, I have this set up to clean out
old forecasts. We'll need to have a discussion about how much historical
data to hold on to, but for now we'll only keep one forecast for a given
forecast time and we'll clean out any forecasts that are more than 6
hours in the past.