Update /xsection/ to read new variable names
I've changed how the cross-section data is generated to try to match the
new variables in the Zarr, but there's a dimension mismatch when I call
log_interpolate_1d that I don't really understand.

It looks like the 'pres' variable was replaced by 'isobaric', and there
is no temperature variable, so I've removed that.