Replace forecast with runHour/validTime
Everywhere we work with state, I've replaced the forecast property with
runTime and validTime, since this is how we're going to select our
forecasts. I suspect I will end up splitting this state apart so that
we can show the vertically integrated smoke when we have a forecast and
no path.