Handle null path store in App.svelte
Updated the popstate handler in the main app for the new stores: runHour
and validTime are separate stores, and path may be null, so we set the
stores from the state accordingly.

I also added some guards to prevent dereferencing properties on a null
path store when we set up our start and end properties.