Handle a null path store in the HrrrMap
I removed the ready property because we can just check of $path is
truthy. I've also changed the conditions for drawing. If we have border
data, we'll draw the map whether or not there's a path. This way we can
draw a map of CONUS with vertically integrated smoke before a path is
selected to help people (me) find coordinates to input for a path, if
they don't have one in mind already.