Display vertically integrated smoke on the map
I'm not sure if I'm handling the projection correctly as I try to map
my contours back from the interpolated grid values to lat, lng
coordinates for mapping, but it looks approximately correct, when I
eyeball it against one of the production graphics.

I think it might be better to figure out a way to generate these
contours on the server and send the GeoJSON for the contours to the
frontend. That would make for a lot smaller payloads in the requests