Update forecast automatically
Bind the forecast settings directly to the runHour and validTime stores
so that we fetch new forecasts as they are selected. This works fine for
these fields because selections are (mostly) atomic, unlike the path
selection where we might trigger a fetch while a person is in the midst
of typing a coordinate.

This means that the vertically integrated smoke map will update more
quickly.