Fix the color scale for the xsection
I had some weird residual scale stuff going on with the colors in the
XSection component. It was using scaleSqrtSequential, which I think was
a holdover from when I was allowing d3 to select its own thresholds.
I've applied the same threshold scale technique (albeit with different
thresholds, for now) that I used in the map. This creates an even
spacing of colors regardless of the thresholds.