Add a title to the legend
Use a grid to arrange the title and axis of the legend more easily. I
hard-coded the width of the swatches so that I can set the main app grid
column width to min-content for the legend. This takes some of the
guesswork out of sizing the legend and ensures that it should mostly
be legible.