Add a loading spinner component
This is the most basic loading indicator: it just says "Loading". But I
can wrap any component in it and provide the promise that we're awaiting
and it will display the loading indicator over the component, so that's
good.

Right now it's set up for the map, going to do the cross-section plot
next.