Fix an infinite loop in the XSection component
Because I was reading from xScale and yScale variables and writing to
them in the same reactive block, the block was scheduled for another
execution every time it executed. I've separated out each of the
variables from the API response and put each reactive assignment in its
own block so that the dependencies of each block are more accurately
determined.