Add layout for stacking components atop one another
I've renamed the stack layout to flow and changed the behavior of the
stack layout to pile children one on top of another (maybe I should've
called it "pile"). This layout is only used by the loading component,
but may come in handy elsewhere.

This stack technique comes from https://smolcss.dev/#smol-stack-layout