Fix drawing order on map
Because I moved to using Promises for detecting loading states, the
smoke was being drawn on top of the path, which often covered it. I
refactored the render() function into a reactive block so that whenever
anything changes, the whole map gets redrawn in the right order. It uses
guards to skip over parts it's missing data for.