Debugging the map and legend resizing
Trying to fix this bug where resizing the window triggers infinite
growth of the main area. I think I've fixed the sizing issue for the map
and the legend, although the map is not redrawing the way it ought. I
can't help but feeling there's a lot of extra divs around here. I wonder
if I could optimize some of this.