Convert the draw code back to a function
The reason the map was clearing an not redrawing was, I think, because
changes to the size of a <canvas> element clear the canvas
automatically. The draw code was running in response to the changing
data before the DOM finished updating, so the updated drawing was
cleared. As a function, I can now call it from afterUpdate to avoid
this issue, and I can trigger it in a reactive block by just listing
the properties I want the block to depend on.