Prevent multiple fetches of the border data
The border data doesn't change, so I moved it into a store and cached
the value so that we don't refetch it every time we go from the CONUS
map to a cross-section and back. The HrrrMap component was getting
destroyed and recreated each time, so the onMount callback was run each
time.