Draw the in-progress path on the map
Add a second canvas for drawing the in-progress path when a user clicks
on the map. The second canvas allows us to draw the path without having
to keep redrawing all of the data on the map. Could be a good idea to
use multiple canvases to make our drawing more efficient…