Move styles for coordinate inputs into the component
First test of the SCSS preprocessing in component styles is successful.
The styles for the coordinate input are now in the Svelte component and
successfully reading the gap size from the USWDS functions. I also
copied the label styles for the uppercase labels into the FormGroup
component, though I think I'll want to create some font tokens as
utility classes and use those instead.