Return GeoJSON from /vertical/
Use matplotlib to generate the contours (it feels silly to define a
figure and a plot just to generate the paths from this data, but c'est
la vie) and then run the contours through geojsoncontour to generate
the response.

The file size is < 1MB which is a *huge* improvement, if this actually
works.