Pass thresholds and colors to backend
We pass the thresholds and colors for the contours to the backend now
so that matplotlib will make use of them. The thresholds are necessary;
we could skip the colors and do that client-side, but because of how
geojsoncontour builds the FeatureCollection, it's easier to tell
matplotlib what colors we want for each level.