Update the HrrrMap to expect GeoJSON
Eliminate all of the code that generated and projected the contours for
the map. We should be able to use the same geopath generator we use for
the boundaries to draw the contours now.

Something is not working quite right, but when I check the GeoJSON from
the API in geojson.io, it looks right, so there is probably something
wrong in my JS.