Use geoTransform to project points
There is something odd about the GeoJSON that I'm getting from combining
matplotlib's contourf plot with contourgeojson. The GeoJSON works fine
when I display it in a mapbox-based tool like https://geojson.io or
Deck.gl, but in D3 I get a polygon that covers the entire globe for all
layers. d3.geoBounds reports [[-180, -90], [180, 90]] for the GeoJSON,
which is not correct, the data only extends a little beyond CONUS,so I
think there's a subtle bug somewhere.

If I transform each point individually instead of letting D3 project the
entire MultiPolygon, everything is drawn correctly. This makes me
suspicious that issue may have to do with the clipping that D3 does for
projections that may not be part of the more generic transformations.

I'd like to track down the issue, but for now, this is working.