Improve contrast of borders against contours on map
I changed the drawing order so that borders are drawn on top of the
contours, rather than below. This eliminates the need for opacity on the
contours, so there's less likely to be an issue caused by blending
colors in the map.

I had to lighten the borders and I shrank the width of the county
borders a little so that they don't interfere with the contours. It's
not perfect, the counties should be more visible in the inset map than
they are. I'm still contemplating how best to handle providing this kind
of context for the map.